Move buttons and markup to src
