+++ main.py
@@ -7,6 +7,7 @@ from telebot.asyncio_storage import StateMemoryStorage
 from telebot.asyncio_handler_backends import State, StatesGroup
 
 import utils.settings as settings, logging
+from src.buttons import get_main_markup, Button
 from utils.db import User, Payment, Notification
 
 logger = logging.getLogger(__name__)
@@ -20,16 +21,11 @@ class ChangeNameStates(StatesGroup):
     change_name = State()
 
 
-class Button:
-    rename = 'Изменить имя'
-    payments = 'Показать список сервисов'
-    add_new_payment = 'Добавить новый сервис по оплате'
-    change_payment = 'Изменить информацию'
-    delete_payment = 'Удалить'
-    notifications = 'Список уведомлений'
-    add_new_notification = 'Добавить уведомление'
-    change_notification = 'Изменить уведомление'
-    delete_notification = 'Удалить'
+class PaymentStates(StatesGroup):
+    payment_list = State()
+    payment_add = State()
+    payment_change = State()
+    payment_delete = State()
 
 
 @bot.message_handler(commands=['help', 'start'])
@@ -71,16 +67,19 @@ async def change_name(message):
     await main_buttons(message.chat.id)
 
 
-
+@bot.message_handler(text_contains=[Button.payments])
+async def payments_list(message):
+    logger.debug(f'User select payments list')
+    await bot.send_message(
+        message.chat.id,
+        'Твой список платежей:'
+    )
+    await bot.set_state(message.from_user.id, PaymentStates.payment_list, message.chat.id)
 
 
 async def main_buttons(chat_id: int) -> None:
-    markup = types.ReplyKeyboardMarkup(row_width=2)
-    user_rename = types.KeyboardButton(Button.rename)
-    payments_list = types.KeyboardButton(Button.payments)
-    markup.add(user_rename, payments_list)
     logger.debug(f'Show user main buttons')
-    await bot.send_message(chat_id, 'Чего изволите?', reply_markup=markup)
+    await bot.send_message(chat_id, 'Чего изволите?', reply_markup=get_main_markup())
 
 
 if __name__ == '__main__':
